+++ src/lib.rs
@@ -39,7 +39,9 @@ pub(crate) mod window_common;
 #[cfg(target_os = "windows")]
 pub use app::PlatformInterface;
 pub use error::{PlatformError, Result as PlatformResult};
-pub use styling_primitives::{Color, ControlStyle, FontDescription, FontWeight, StyleId};
+pub use styling_primitives::{
+    Color, ControlStyle, FontDescription, FontWeight, StyleId, TextAlignment,
+};
 pub use types::{
     AppEvent, BadgeDescriptor, ChartDataPacket, ChartLineData, ChartLineEmphasis, CheckState,
     ControlId, DockStyle, FormButtons, FormDialogDescriptor, FormField, FormFieldValue,